Update 'workspace/stm32fxx_drivers/' - Update I2C peripheral driver, create I2C application 1

+++ workspace/stm32fxx_drivers/Drivers/Inc/stm32f407xx_i2c_driver.h
@@ -17,7 +17,7 @@
 /* I2Cx peripheral configuration structure */
 typedef struct
 {
-	uint8_t I2C_SCLSpeed;		/* Available values @I2C_SCLSpeed	 		*/
+	uint32_t I2C_SCLSpeed;		/* Available values @I2C_SCLSpeed	 		*/
 	uint8_t I2C_DeviceAddress;	/* Value will be entered by user			*/
 	uint8_t I2C_ACKEnable;		/* Available values @I2C_ACKEnable	 		*/
 	uint8_t I2C_FMDutyCycle;	/* Available values @I2C_FMDutyCycle 		*/
